requirements.txt
aiogram==3.6.0
aiosqlite
python-dotenv